backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
datasets==2.12.0
dill==0.3.6
dnspython==2.3.0
dotenv-python==0.0.1
filelock==3.12.0
Flask==2.3.2
Flask-Cors==3.0.10
flaskwebgui==1.0.6
frozenlist==1.3.3
fsspec==2023.5.0
googleapis-common-protos==1.56.4
greenlet==2.0.2
grpc-gateway-protoc-gen-openapiv2==0.1.0
grpcio==1.54.2
huggingface-hub==0.14.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.162
loguru==0.7.0
lz4==4.3.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.7
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.1
pinecone-client==2.2.1
protobuf==3.19.3
psutil==5.9.5
pyarrow==12.0.0
pydantic==1.10.8
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
responses==0.18.0
six==1.16.0
SQLAlchemy==2.0.15
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.1
tzdata==2023.3
urllib3==2.0.2
Werkzeug==2.3.4
xxhash==3.2.0
yarl==1.9.2
